# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Transport Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /hubs:
    get:
      operationId: listHubs
      tags: [hubs]
      summary: List all hubs
      responses:
        "200":
          description: List of hubs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Hub"

    post:
      operationId: createHub
      tags: [hubs]
      summary: Create a hub
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HubInput"
      responses:
        "201":
          description: Hub created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hub"
  /hubs/{hubId}:
    get:
      operationId: getHub
      tags: [hubs]
      summary: Get hub by ID
      parameters:
        - name: hubId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hub
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hub"

    patch:
      operationId: updateHub
      tags: [hubs]
      summary: Update a hub
      parameters:
        - name: hubId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HubUpdate"
      responses:
        "200":
          description: Hub updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hub"
  /hubs/{hubId}/toggle-active:
    patch:
      operationId: toggleHubActive
      tags: [hubs]
      summary: Toggle hub active status
      parameters:
        - name: hubId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hub status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hub"

  # STAFF
...
  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      summary: List complaints
      parameters:
        - name: hubId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: parcelId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"

    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Raise a complaint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintInput"
      responses:
        "201":
          description: Complaint created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
  /complaints/{complaintId}:
    get:
      operationId: getComplaint
      tags: [complaints]
      summary: Get complaint details
      parameters:
        - name: complaintId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Complaint
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"

    patch:
      operationId: updateComplaint
      tags: [complaints]
      summary: Update complaint (assign, resolve, close)
      parameters:
        - name: complaintId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComplaintUpdate"
      responses:
        "200":
          description: Complaint updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"

  # REPORTS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, staff]
      properties:
        token:
          type: string
        staff:
          $ref: "#/components/schemas/StaffMember"
    Hub:
      type: object
      required: [id, hubName, hubCode, address, contactNumber, isActive, createdAt]
      properties:
        id:
          type: integer
        hubName:
          type: string
        hubCode:
          type: string
        address:
          type: string
        contactNumber:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    HubInput:
      type: object
      required: [hubName, hubCode, address, contactNumber]
      properties:
        hubName:
          type: string
        hubCode:
          type: string
        address:
          type: string
        contactNumber:
          type: string
    HubUpdate:
      type: object
      properties:
        hubName:
          type: string
        hubCode:
          type: string
        address:
          type: string
        contactNumber:
          type: string
    StaffMember:
      type: object
      required: [id, name, phone, email, role, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
        hubId:
          type: ["integer", "null"]
        hubName:
          type: ["string", "null"]
        hubCode:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    Complaint:
      type: object
      required: [id, complaintNumber, parcelId, raisedBy, complaintType, description, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
        complaintNumber:
          type: string
        parcelId:
          type: integer
        awbNumber:
          type: ["string", "null"]
        raisedBy:
          type: integer
        raisedByName:
          type: ["string", "null"]
        complaintType:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
        status:
          type: string
        assignedTo:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        resolutionNotes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    ComplaintInput:
      type: object
      required: [parcelId, complaintType, description]
      properties:
        parcelId:
          type: integer
        complaintType:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
    ComplaintUpdate:
      type: object
      properties:
        status:
          type: string
        assignedTo:
          type: ["integer", "null"]
        resolutionNotes:
          type: string
...